Release 4.0.0
Importing release 4.0.0

+++ src/libs/zbxnix/daemon.c
@@ -23,7 +23,7 @@
 #include "pid.h"
 #include "cfg.h"
 #include "log.h"
-#include "zbxself.h"
+#include "control.h"
 
 #include "fatal.h"
 #include "sighandler.h"
@@ -196,7 +196,7 @@ static void	user1_signal_handler(int sig, siginfo_t *siginfo, void *context)
 			SIG_CHECKED_FIELD(siginfo, si_pid),
 			SIG_CHECKED_FIELD(siginfo, si_uid),
 			SIG_CHECKED_FIELD(siginfo, si_value.ZBX_SIVAL_INT),
-			SIG_CHECKED_FIELD(siginfo, si_value.ZBX_SIVAL_INT));
+			(unsigned int)SIG_CHECKED_FIELD(siginfo, si_value.ZBX_SIVAL_INT));
 #ifdef HAVE_SIGQUEUE
 	flags = SIG_CHECKED_FIELD(siginfo, si_value.ZBX_SIVAL_INT);
